Changement int => double

+++ Dossier-General/Programme-Principal/src/cartes/CarteDouble.java
@@ -4,20 +4,20 @@ import elements.Colonne;
 
 public abstract class CarteDouble extends CarteInfluence {
 
-	private int valeurSpéciale; // Valeur spéciale utilisée lorsque que la carte est placée dans la colonne d'une carte objectif ayant le meme domaine
+	private double valeurSpéciale; // Valeur spéciale utilisée lorsque que la carte est placée dans la colonne d'une carte objectif ayant le meme domaine
 
 	private String domaine;
 
 	
 	// Constructeur basique
-	protected CarteDouble(String nom, int valeur, boolean visible, Colonne colonne, int valeurSpéciale, String domaine) {
+	protected CarteDouble(String nom, double valeur, boolean visible, Colonne colonne, double valeurSpéciale, String domaine) {
 		super(nom, valeur, visible, colonne);
 		this.valeurSpéciale = valeurSpéciale;
 		this.domaine = domaine;
 	}
 
 	// Constructeur simplifié
-	protected CarteDouble(String nom, int valeur, int valeurSpéciale, String domaine) {
+	protected CarteDouble(String nom, double valeur, double valeurSpéciale, String domaine) {
 		super(nom, valeur, false, null);
 		this.valeurSpéciale = valeurSpéciale;
 		this.domaine = domaine;
@@ -30,11 +30,11 @@ public abstract class CarteDouble extends CarteInfluence {
 		this.domaine = domaine;
 	}
 
-	public int getValeurSpé() {
+	public double getValeurSpé() {
 		return valeurSpéciale;
 	}
 
-	public void setValeurSpé(int valeurSpéciale) {
+	public void setValeurSpé(double valeurSpéciale) {
 		this.valeurSpéciale = valeurSpéciale;
 	}
 
@@ -49,7 +49,7 @@ public abstract class CarteDouble extends CarteInfluence {
 	
 	//Opération
 	
-	public int ValeurUtilisee(CarteObjectif objectif) {
+	public double ValeurUtilisee(CarteObjectif objectif) {
 		if (this.domaine == objectif.getDomaine()) {
 			return this.valeurSpéciale;
 		}
